User模块已完成开发（使用的是jar包调用api的模式，因此更改了配置文件，使用时需注意jar包的依赖使用的是绝对路径，因此运行测试时需更改） 注意： RestTemplateConfig文件配置有冲突，待解决

+++ src/main/java/com/miniBOM/pojo/User.java
@@ -6,7 +6,7 @@ import net.minidev.json.annotate.JsonIgnore;
 
 @Data
 public class User {
-    private Integer id;//->ID
+    private Long id;//->ID
     private String name;//->Name
     @JsonIgnore
     private String password;//->UserPassword
